pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "spcrawler"
version = "1.0.0"
description = "Sports piracy stream detection crawler"
readme = "README.MD"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.4.0",
    "ddgs>=6.0.0",
    "pymongo>=4.6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = ["spcrawler", "spcrawler.client", "spcrawler.instance", "spcrawler.utils"]

[tool.setuptools.package-dir]
spcrawler = "src"
